Minimize required code
